ppc64: support for > 32 CPUs (24 way RS64 with HMT shows up as 48 way)

+++ include/asm-ppc64/smp.h
@@ -52,7 +52,7 @@ static inline int num_online_cpus(void)
 	return nr;
 }
 
-extern volatile unsigned long cpu_callin_map[NR_CPUS];
+extern volatile unsigned int cpu_callin_map[NR_CPUS];
 
 #define smp_processor_id() (get_paca()->xPacaIndex)
